Adicionado suporte ao Delphi 7, solicitado por Renan Teixeira <renan.r.teixeira@gmail.com>.

+++ source/pUCFrame_Profile.pas
@@ -5,9 +5,10 @@ interface
 {$I 'UserControl.inc'}
 
 uses
-  {$IFDEF DELPHI5_UP}
+ {$IFDEF DELPHI5_UP}
+ {$ELSE}
   Variants,
-  {$ENDIF}
+ {$ENDIF}
   Buttons,
   Classes,
   Controls,
@@ -23,22 +24,18 @@ uses
   StdCtrls,
   SysUtils,
   UcBase,
-  UserPermis_U
-  {$IFNDEF FPC},
-  Windows
-  {$ELSE},
-  LCLType
-  {$ENDIF};
+  UserPermis_U,
+  Windows;
 
 type
   TFrame_Profile = class(TFrame)
     DbGridPerf: TDBGrid;
-    Panel2: TPanel;
-    BtnAddPer: TBitBtn;
-    BtnAltPer: TBitBtn;
-    BtnExcPer: TBitBtn;
-    BtnClose: TBitBtn;
-    BtnAcePer: TBitBtn;
+    Panel2:     TPanel;
+    BtnAddPer:  TBitBtn;
+    BtnAltPer:  TBitBtn;
+    BtnExcPer:  TBitBtn;
+    BtnClose:   TBitBtn;
+    BtnAcePer:  TBitBtn;
     DataPerfil: TDataSource;
     procedure BtnAddPerClick(Sender: TObject);
     procedure BtnAltPerClick(Sender: TObject);
@@ -52,11 +49,12 @@ type
   private
     { Private declarations }
   public
-    fUserControl: TUserControl;
+    fUserControl:          TUserControl;
     FDataSetPerfilUsuario: TDataset;
-    destructor Destroy; override;
+    destructor Destroy; override;    
     { Public declarations }
   end;
+
 
 implementation
 
@@ -72,10 +70,10 @@ begin
     else
       FIncluirPerfil.LbDescricao.Caption := LabelChange;
 
-    FIncluirPerfil.lbNome.Caption := LabelName;
-    FIncluirPerfil.btGravar.Caption := BtSave;
-    FIncluirPerfil.btCancela.Caption := BtCancel;
-    FIncluirPerfil.Position := fUserControl.UserSettings.WindowsPosition;
+    FIncluirPerfil.lbNome.Caption        := LabelName;
+    FIncluirPerfil.btGravar.Caption      := BtSave;
+    FIncluirPerfil.btCancela.Caption     := BtCancel;
+    FIncluirPerfil.Position              := FUserControl.UserSettings.WindowsPosition;
     FIncluirPerfil.FDataSetPerfilUsuario := FDataSetPerfilUsuario;
   end;
 end;
@@ -84,30 +82,33 @@ procedure TFrame_Profile.ActionBtPermissProfileDefault;
 var
   TempCampos, TempCamposEX: String;
 begin
-  UserPermis.FTempIdUser := FDataSetPerfilUsuario.FieldByName('IdUser')
-    .AsInteger;
+  UserPermis.FTempIdUser := FDataSetPerfilUsuario.FieldByName('IdUser').AsInteger;
   with fUserControl do
   begin
-    TempCampos :=
-      Format(' %s as IdUser, %s as Modulo, %s as ObjName, %s as UCKey ',
-      [TableRights.FieldUserID, TableRights.FieldModule,
-      TableRights.FieldComponentName, TableRights.FieldKey]);
-    TempCamposEX := Format('%s, %s as FormName ',
-      [TempCampos, TableRights.FieldFormName]);
+    TempCampos   := Format(' %s as IdUser, %s as Modulo, %s as ObjName, %s as UCKey ',
+      [TableRights.FieldUserID,
+      TableRights.FieldModule,
+      TableRights.FieldComponentName,
+      TableRights.FieldKey]);
+    TempCamposEX := Format('%s, %s as FormName ', [TempCampos, TableRights.FieldFormName]);
 
-    UserPermis.DSPermiss := DataConnector.UCGetSQLDataset
-      (Format('Select %s from %s tab Where tab.%s = %s and tab.%s = %s',
-      [TempCampos, TableRights.TableName, TableRights.FieldUserID,
+    UserPermis.DSPermiss := DataConnector.UCGetSQLDataset(Format('Select %s from %s tab Where tab.%s = %s and tab.%s = %s',
+      [TempCampos,
+      TableRights.TableName,
+      TableRights.FieldUserID,
       FDataSetPerfilUsuario.FieldByName('IdUser').AsString,
-      TableRights.FieldModule, QuotedStr(ApplicationID)]));
+      TableRights.FieldModule,
+      QuotedStr(ApplicationID)]));
 
     UserPermis.DSPermiss.Open;
 
-    UserPermis.DSPermissEX := DataConnector.UCGetSQLDataset
-      (Format('Select %s from %s tab1 Where tab1.%s = %s and tab1.%s = %s',
-      [TempCamposEX, TableRights.TableName + 'EX', TableRights.FieldUserID,
+    UserPermis.DSPermissEX := DataConnector.UCGetSQLDataset(Format('Select %s from %s tab1 Where tab1.%s = %s and tab1.%s = %s',
+      [TempCamposEX,
+      TableRights.TableName + 'EX',
+      TableRights.FieldUserID,
       FDataSetPerfilUsuario.FieldByName('IdUser').AsString,
-      TableRights.FieldModule, QuotedStr(ApplicationID)]));
+      TableRights.FieldModule,
+      QuotedStr(ApplicationID)]));
 
     UserPermis.DSPermissEX.Open;
 
@@ -127,18 +128,17 @@ procedure TFrame_Profile.SetWindowProfile;
 begin
   with fUserControl.UserSettings.Rights do
   begin
-    UserPermis.Caption := WindowCaption;
-    UserPermis.LbDescricao.Caption := LabelProfile;
-    UserPermis.lbUser.Left := UserPermis.LbDescricao.Left +
-      UserPermis.LbDescricao.Width + 5;
-    UserPermis.PageMenu.Caption := PageMenu;
-    UserPermis.PageAction.Caption := PageActions;
+    UserPermis.Caption              := WindowCaption;
+    UserPermis.LbDescricao.Caption  := LabelProfile;
+    UserPermis.lbUser.Left          := UserPermis.LbDescricao.Left + UserPermis.LbDescricao.Width + 5;
+    UserPermis.PageMenu.Caption     := PageMenu;
+    UserPermis.PageAction.Caption   := PageActions;
     UserPermis.PageControls.Caption := PageControls; // By Vicente Barros Leonel
-    UserPermis.BtLibera.Caption := BtUnlock;
-    UserPermis.BtBloqueia.Caption := BtLock;
-    UserPermis.BtGrava.Caption := BtSave;
-    UserPermis.BtCancel.Caption := BtCancel;
-    UserPermis.Position := fUserControl.UserSettings.WindowsPosition;
+    UserPermis.BtLibera.Caption     := BtUnlock;
+    UserPermis.BtBloqueia.Caption   := BtLock;
+    UserPermis.BtGrava.Caption      := BtSave;
+    UserPermis.BtCancel.Caption     := BtCancel;
+    UserPermis.Position             := fUsercontrol.UserSettings.WindowsPosition;
   end;
 end;
 
@@ -146,19 +146,18 @@ procedure TFrame_Profile.BtnAcePerClick(Sender: TObject);
 begin
   if FDataSetPerfilUsuario.IsEmpty then
     Exit;
-  UserPermis := TUserPermis.Create(self);
-  UserPermis.fUserControl := fUserControl;
+  UserPermis              := TUserPermis.Create(self);
+  UserPermis.FUserControl := fUsercontrol;
   SetWindowProfile;
-  UserPermis.lbUser.Caption := FDataSetPerfilUsuario.FieldByName
-    ('Nome').AsString;
+  UserPermis.lbUser.Caption := FDataSetPerfilUsuario.FieldByName('Nome').AsString;
   ActionBtPermissProfileDefault;
 end;
 
 procedure TFrame_Profile.BtnAddPerClick(Sender: TObject);
 begin
   try
-    FIncluirPerfil := TfrmIncluirPerfil.Create(self);
-    FIncluirPerfil.fUserControl := self.fUserControl;
+    FIncluirPerfil              := TfrmIncluirPerfil.Create(Self);
+    FIncluirPerfil.FUserControl := Self.FUserControl;
     SetWindowPerfil(True);
     FIncluirPerfil.ShowModal;
   finally
@@ -171,15 +170,14 @@ begin
   if FDataSetPerfilUsuario.IsEmpty then
     Exit;
   try
-    FIncluirPerfil := TfrmIncluirPerfil.Create(self);
-    FIncluirPerfil.fUserControl := self.fUserControl;
-    FIncluirPerfil.FNewIdUser := FDataSetPerfilUsuario.FieldByName('IdUser')
-      .AsInteger;
+    FIncluirPerfil              := TfrmIncluirPerfil.Create(self);
+    FIncluirPerfil.FUserControl := Self.FUserControl;
+    FIncluirPerfil.FNewIdUser   := FDataSetPerfilUsuario.FieldByName('IdUser').AsInteger;
     SetWindowPerfil(False);
     with FIncluirPerfil do
     begin
       EditDescricao.Text := FDataSetPerfilUsuario.FieldByName('Nome').AsString;
-      FAltera := True;
+      FAltera            := True;
       ShowModal;
     end;
   finally
@@ -189,38 +187,26 @@ end;
 
 procedure TFrame_Profile.BtnExcPerClick(Sender: TObject);
 var
-  TempID: Integer;
+  TempID:    Integer;
   CanDelete: Boolean;
-  ErrorMsg: String;
-  TempDS: TDataset;
+  ErrorMsg:  String;
+  TempDS:    TDataset;
 begin
   if FDataSetPerfilUsuario.IsEmpty then
     Exit;
   TempID := FDataSetPerfilUsuario.FieldByName('IDUser').AsInteger;
-  TempDS := fUserControl.DataConnector.UCGetSQLDataset
-    ('Select ' + fUserControl.TableUsers.FieldUserID + ' as IdUser from ' +
-    fUserControl.TableUsers.TableName + ' Where ' +
-    fUserControl.TableUsers.FieldTypeRec + ' = ' + QuotedStr('U') + ' AND ' +
-    fUserControl.TableUsers.FieldProfile + ' = ' + IntToStr(TempID));
+  TempDS := FUserControl.DataConnector.UCGetSQLDataset('Select ' + FUserControl.TableUsers.FieldUserID + ' as IdUser from ' +
+    FUserControl.TableUsers.TableName +
+    ' Where ' + FUserControl.TableUsers.FieldTypeRec + ' = ' + QuotedStr('U') +
+    ' AND ' + FUserControl.TableUsers.FieldProfile + ' = ' + IntToStr(TempID));
 
   if TempDS.FieldByName('IdUser').AsInteger > 0 then
   begin
     TempDS.Close;
     FreeAndNil(TempDS);
-    // changed by fduenas: PromptDelete_WindowCaption
-    {$IFNDEF FPC}
-    if MessageBox(handle,
-      PChar(Format(fUserControl.UserSettings.UsersProfile.PromptDelete,
-      [FDataSetPerfilUsuario.FieldByName('Nome').AsString])),
-      PChar(fUserControl.UserSettings.UsersProfile.PromptDelete_WindowCaption),
-      MB_ICONQUESTION or MB_YESNO or MB_DEFBUTTON2) <> idYes then
-    {$ELSE}
-    if Application.MessageBox(
-      PChar(Format(fUserControl.UserSettings.UsersProfile.PromptDelete,
-      [FDataSetPerfilUsuario.FieldByName('Nome').AsString])),
-      PChar(fUserControl.UserSettings.UsersProfile.PromptDelete_WindowCaption),
-      MB_ICONQUESTION + MB_YESNO) <> mrYes then
-    {$ENDIF}
+    //changed by fduenas: PromptDelete_WindowCaption
+    if MessageBox(handle, PChar(Format(FUserControl.UserSettings.UsersProfile.PromptDelete, [FDataSetPerfilUsuario.FieldByName('Nome').AsString])),
+      PChar(FUserControl.UserSettings.UsersProfile.PromptDelete_WindowCaption), MB_ICONQUESTION or MB_YESNO or MB_DEFBUTTON2) <> idYes then
       Exit;
   end
   else
@@ -230,25 +216,21 @@ begin
   end;
 
   CanDelete := True;
-  if Assigned(fUserControl.onDeleteProfile) then
-    fUserControl.onDeleteProfile(TObject(Owner), TempID, CanDelete, ErrorMsg);
+  if Assigned(FUserControl.onDeleteProfile) then
+    FUserControl.onDeleteProfile(TObject(Owner), TempID, CanDelete, ErrorMsg);
   if not CanDelete then
   begin
-    MessageDlg(ErrorMsg, mtWarning, [mbOK], 0);
+    MessageDlg(ErrorMSG, mtWarning, [mbOK], 0);
     Exit;
   end;
 
-  with fUserControl do
+  with FUserControl do
   begin
-    DataConnector.UCExecSQL('Delete from ' + TableUsers.TableName + ' where ' +
-      TableUsers.FieldUserID + ' = ' + IntToStr(TempID));
-    DataConnector.UCExecSQL('Delete from ' + TableRights.TableName + ' where ' +
-      TableRights.FieldUserID + ' = ' + IntToStr(TempID));
-    DataConnector.UCExecSQL('Delete from ' + TableRights.TableName + 'EX where '
-      + TableRights.FieldUserID + ' = ' + IntToStr(TempID));
-    DataConnector.UCExecSQL('Update ' + TableUsers.TableName + ' Set ' +
-      TableUsers.FieldProfile + ' = null where ' + TableUsers.FieldUserID +
-      ' = ' + IntToStr(TempID));
+    DataConnector.UCExecSQL('Delete from ' + TableUsers.TableName + ' where ' + TableUsers.FieldUserID + ' = ' + IntToStr(TempID));
+    DataConnector.UCExecSQL('Delete from ' + TableRights.TableName + ' where ' + TableRights.FieldUserID + ' = ' + IntToStr(TempID));
+    DataConnector.UCExecSQL('Delete from ' + TableRights.TableName + 'EX where ' + TableRights.FieldUserID + ' = ' + IntToStr(TempID));
+    DataConnector.UCExecSQL('Update ' + TableUsers.TableName +
+      ' Set ' + TableUsers.FieldProfile + ' = null where ' + TableUsers.FieldUserID + ' = ' + IntToStr(TempID));
   end;
   FDataSetPerfilUsuario.Close;
   FDataSetPerfilUsuario.Open;
@@ -256,9 +238,9 @@ end;
 
 destructor TFrame_Profile.Destroy;
 begin
-  // nada a destruir
-  // não destruir o FDataSetPerfilUsuario o USERCONTROL toma conta dele
+  //nada a destruir
+  //não destruir o FDataSetPerfilUsuario o USERCONTROL toma conta dele
   inherited;
 end;
 
-end.
+end.
